+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 0.0.7 - in progress
 - Support iteration over result sets.
 - Add python environments to gitignore.
+- Updating doc tests, adding new query type
+- Add new parameter to enable gene->dataset and protein->dataset queries
+- Running tests on client test server
 
 0.0.6 - 2021-03-24
 - Parameterize api endpoint doctests.